split parse into submodules
