+++ packages/stylist-macros/src/inline/output/sheet.rs
@@ -1,7 +1,6 @@
 use super::Reify;
-use proc_macro2::{Span, TokenStream};
+use proc_macro2::TokenStream;
 use quote::quote;
-use syn::Ident;
 
 pub struct OutputSheet {
     pub contents: Vec<TokenStream>,
@@ -9,15 +8,16 @@ pub struct OutputSheet {
 
 impl Reify for OutputSheet {
     fn into_token_stream(self) -> TokenStream {
-        let ident_scopes = Ident::new("scopes", Span::mixed_site());
         let Self { contents } = self;
 
         quote! {
             {
-                let #ident_scopes: ::std::vec::Vec::<::stylist::ast::ScopeContent> = ::std::vec![
-                    #( #contents, )*
-                ];
-                ::stylist::ast::Sheet::from(#ident_scopes)
+                use ::std::convert::{From, Into};
+                ::stylist::ast::Sheet::from(
+                    ::std::vec![
+                        #( ::stylist::ast::ScopeContent::from(#contents), )*
+                    ]
+                )
             }
         }
     }
